feat(api): split abuse-prevention — baseline → core, graduated response → /ee (WS5)
Split abuse-prevention along the documented enterprise boundary
(docs/development/enterprise-gating.md, which lists abuse-prevention as a
/ee feature):

- BASELINE anomaly detection STAYS in core: the sliding-window threshold
  config + parsers, enum-drift coercion, and counter sanitization now live
  in lib/security/abuse-baseline.ts. A security primitive every deployment
  gets.
- The MULTI-TENANT graduated warn→throttle→suspend RESPONSE engine + admin
  operations MOVED to ee/src/abuse-prevention/ (engine.ts), reached through
  a new AbuseResponse Context.Tag (Effect routes) and a sync
  AbuseResponsePolicy holder (non-Effect hot-path call sites). Core ships an
  inert NOOP_ABUSE_RESPONSE_POLICY + NoopAbuseResponseLayer so non-enterprise
  behavior is unchanged.

lib/security/abuse.ts is now a thin delegating shim keeping every historical
import path + signature stable (audit.ts, agent-gate.ts, admin-orgs.ts,
platform-admin.ts, layers.ts, migrate.ts need no edits). admin-abuse.ts
routes yield* AbuseResponse and 404 not_available when EE is absent (SLA
precedent). EELayer/EnterpriseSubsystem/NoopEnterpriseDefaultsLayer gain
AbuseResponse.

Behavior on an enterprise deployment is identical: EE registers the engine
into the Tag + sync holder at layer construction, so the graduated response
runs exactly as before. The core→/ee inversion invariant holds
(check-ee-imports.sh green). The engine-behavior test moved to
ee/src/abuse-prevention/engine.test.ts.

Closes #4000
